proper messages/actions for log cleanup

+++ inc/logging.sh
@@ -17,9 +17,13 @@ log() {
 
 cleanupLogs() {
     REMOVE=$(expr $(ls $LOGDIR/ -1rt | wc -l) - $LOGS_TO_KEEP)
-    for i in $(ls $LOGDIR/ -1rt | head -n $REMOVE);
-    do
-        rm "$LOGDIR/$i";
-    done
-    log "Cleaned up $REMOVE old log files, $LOGS_TO_KEEP remain"
+    if [ "$REMOVE" > 1]; then
+        for i in $(ls $LOGDIR/ -1rt | head -n $REMOVE);
+        do
+            rm "$LOGDIR/$i";
+        done
+        log "Cleaned up $REMOVE old log files,  remain"
+    else
+        log "Fewer then $LOGS_TO_KEEP log fils remain, all retained."
+    fi
 }
